Fix style errors
